CHRE: fix DT_RELA table relocation for Tinysys dynamic loader
This fixes the "Unsupported relocation type 0" issue we see for
some of the Tinysys CHQTS nano apps which happens when the ".rela.dyn"
section header has different sh_addr and sh_offset values. We were
using the sh_addr when looking for the RELA entries, but actually
need to use the sh_offset.

Bug: 268675229
Test: build, flash, run CHQTS nano apps
Change-Id: I59bd7acecddecd2d3c9253c4344ce2bb5c73d2bc