Remove EventLoopManager::deferCallback check in SensorRequestManager
This partially reverts cf7d88dca38aaf912f4c29ae116c4b6b3f2217e0.

For an on-change sensor where deferCallback fails in updateLastEvent,
we still free the event when the data is processed by the event loop
in the else-statement at the end of handleSensorDataEvent. This
data will be freed here. With the previous change, this data would
be freed twice.

Bug: 269259493
Test: Presubmits
Change-Id: If8f8a7e9781b1319daacf76e9452f2dad50a94a5