Remove usage of manually unloading all nanoapps in system/chre tests
We now use enableTestMode() and disableTestMode(), so we do not
need to unload all nanoapps to run a test. We gate the older
unloading-all-nanoapps code to device that are less than
version U so that the tests will still pass. On newer devices,
test mode will handle unloading all nanoapps.

Bug: 266099804
Test: Run CHQTS
Change-Id: Ib293f6d529cab5768ade0228f7c313d43136751e