Move GtsContextHubHostTestUtil to system/chre
This CL moves GtsContextHubHostTestUtil to system/chre so it can
be used by normal CHQTS tests as well as multi-device tests. This
CL does not change the content of GtsContextHubHostTestUtil.java,
except for a small amount of cleanup and a rename to drop the Gts
prefix. Future CLs will combine this file with other test utilities
in a larger cleanup effort.

This CL also removes obsolete TODOs.

Bug: 277626129
Test: make gts
Change-Id: Ib5998dc90854866cc0ad2ae3c912e6132f2ebc73