Use nanoseconds as time in Task and TaskManager
This helps reduce flakiness when we request ns in linux PALs.
This reduces flakiness by reducing rounding errors from requests
that are < 1 ms. When these kind of reqeusts are made (see the
audio PAL), the duration_cast will round down, causing the
time to be effectively 0 ms, which results in strange TaskManager
behavior due to how it treats repetition.

Bug: 288037919
Test: Run chre_unit_tests
Test: Run chre_simulation_tests
Test: Run TestBase.AudioCanSubscribeAndUnsubscribeToDataEvents: failed 1/(10^4)
Change-Id: Ie0f08a60c74f1ab76d04e3b95950c1f5219ef4ef