Cleanup audio and BLE impl tests
This CL cleans up a memory leak in the BLE PAL impl test
as well as a temporary fix in the audio PAL impl test.
The temporary fix was to increase the wait time for
receiving an audio data event to reduce flakiness.
This CL's fix should fix the underlying issue of the
flakiness, so the temporary wait time update fix is
reverted.

Bug: 288037919
Test: Run chre_simulation_tests
Test: Run chre_unit_tests
Change-Id: I0639dcb487076b1780f4146026f17250c7dae590