Fix failing PalBleTest.FilteredScan in chre_unit_tests
This CL fixes the failing test PalBleTest.FilteredScan by:

1. Updating the linux BLE PAL to call the callback upon scan
start or stop in the TaskManager thread instead of the execution
thread. This ensures the proper ordering of returning and callback
execution.
2. Updating the test to lock the mutex before starting the scan.
This stops a race condition and ensures that the test thread will
properly wait for the callback thread to notify the condition
variable instead of notifying before the condition variable wait.

This CL also contains:

1. Changing an EXPECT to an ASSERT where we check for a value and
depend on the value existing later in the test.
2. Increase the speed of the test by scanning aggressively.

Bug: 297857275
Test: chre_unit_tests --gtest_filter=*PalBleTest*
Test: --gtest_repeat=1000 # pass
Change-Id: Ia83eb262638dbee72512379228cf9a9eba2ce2c5