Fix handling of request timeout
Context:

The worker thread (chppWorkThreadStart)
1. calls chppTransportGetTimeUntilNextDoWorkNs to know when to schedule
   the next work
2. calls chppWorkThreadHandleSignal to execute the scheduled work

chppWorkThreadHandleSignal is executed when there is work scheduled.

Bug:

When a request timeout, chppTransportGetTimeUntilNextDoWorkNs returns
CHPP_TRANSPORT_TIMEOUT_IMMEDIATE to schedule work immediately.

chppWorkThreadHandleSignal wouldn't call chppTransportDoWork to send the
timeout response. So the request stays in the tiemout state.

On the next iteration of chppWorkThreadStart the same work is scheduled
again, and the request stay in timeout again, ...

Eventually some other work is scheduled and chppTransportDoWork is
called. At that point the timeout response is sent.

But:
- The timeout response should have been sent earlier
- The worker thread keeps looping until chppTransportDoWork is finally
  called

Bug: 266700275
Test: atest chre_chpp_linux_tests
Change-Id: I493576a76b21daa8e5e4f7a1270dd8ff84679462