Fix run_test.sh failing
This is needed since a CL added a header to a target that is built
during unit test. However, we don't invoke CHRE Makefile based unit test
during presubmit or preupload so we didn't catch this error in the
beginning. This CL added ./run_tests.sh -b to preupload to prevent this
from happening.

Bug: 299851035
Test: ./run_tests.sh, ./run_tests.sh -b, ./run_tests.sh --
--gtest_filter=DynamicVector.*

Change-Id: I508a873c2dc37ca467b6b191aa2aa2778efa169d