Add host message counter to debug
Tracking nanoapp message volume can be used to help debug power issues.
Currently we track the number of times a nanoapp is responsible for
waking up the AP. This is good, but does not highlight cases where the
AP is kept awake by a large stream of nanoapp->host messages.

This adds a bucketed counting system for nanoapp->host messages
per-nanoapp, in the same scheme as the 'wakeup' tracking.

Bug: 307816492
Test: ./load_aoc.sh, adb bugreport
Change-Id: Ic937991af902037e57cf6f42f98778d9bc33d87f