Remove unnecessary locking from chre::TaskManager destructor
This CL does the following:

1. Adds the predicate check for mContinueRunningThread to all
condition variables. This ensures that is the flag is set to
false, the worker thread will not wait, but return.
2. Add a check for mContinueRunningThread to the while-loop
in the worker thread. This ensures an earlier return than
checking during execution.
3. Remove the lock in the destructor. If notify_all() is
called before the worker thread locks, then the condition
variable predicates will ensure no waiting will occur.
If the worker thread is waiting, the notify_all() call
will ensure the wait is interrupted and the predicate
will ensure the waits return and the worker thread
returns.
4. Change mContinueRunningThread to an atomic boolean.
5. Cleanup the worker thread run() method.
6. Change Task::execute to calculate execution timestamps
without drift.
7. Only notify the worker thread if the added task is
at the top of the queue.
8. Add TaskManager tests to test whether the
TaskManager can do nothing and whether tasks that are
queued out of order of execution time can be executed
in the correct order.

Bug: 298459533
Test: Run chre_unit_tests with *Task* filter
Test: 1000 iter, all PASS
Change-Id: I97c24b698a5bfaa6d63f456fba594477d866f9f9