Add embedded aconfig flag support to CHRE
This CL adds support to use the aconfig commandline tool to generate
embedded-friendly code flags. These are build-time only flags which will
be optimized out by the compiler.

Flag values can be layered and overwritten. Flags (and associated
artifacts) are re-generated for each required build and should not be committed

The aconfig commandline binary has been cloned into this repo as an interim solution. This will be removed when the aconfig repo is added to all necessary manifests

Bug: 318367982
Test: Add flag dependency, ./load_aoc.sh, ./load_app.sh (Drop)
Change-Id: I01f9856befb8930dc16278189c69dccb68fcd00f