Use CHRE_FIRST_SUPPORTED_API_VERSION to filter unnecessary NSL stubs.
Adds checks of the form around most stubs in NSL:
bool chreGetAudioSource(...) {}
where CHRE_API_VERSION_1_2 introduced chreGetAudioSource().

The stubs for chreBleStartScanAsync(), chreWifiRequestScanAsync(),
chreGetNanoappInfoByAppId(), and chreGetNanoappInfoByInstanceId()
provide backward compatibility for CHRE versions after the versions that
introduced the APIs. As such, the check is against the version after
which that behavior is no longer relevant.

Bug: 239180182
Test: Verified that an NSL stub is excluded from nanoapps when
CHRE_FIRST_SUPPORTED_API_VERSION is at least the version in wrapping #if
statement. Manually verified that the versions used match both the
documentation and stub behavior.

Change-Id: I343ff3dd86742bef9220200f55d19855294747f1